requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
numpy==2.1.2
pandas==2.2.3
pyarrow==16.1.0
scikit-learn==1.5.2
lightgbm==4.5.0
joblib==1.4.2
onnx==1.16.2
onnxruntime==1.19.2
skl2onnx==1.17.0
pyyaml==6.0.2
pytest==8.3.3
httpx==0.27.2
